requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.7
django-bootstrap-v5==1.0.11
Pillow==10.1.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
factory_boy